Moved the ``build_queryset`` method to ``SearchIndex``.
This method is used to build the queryset for indexing operations. It is copied
from the build_queryset function that lived in the update_index management
command.

Making this change allows developers to modify the queryset used for indexing
even when a date filter is necessary. See `tests/core/indexes.py` for tests.